JAVA-41259 Moved code of article jpa-unique-constraints from java-jpa-3 to java-jpa
