+++ persistence-modules/java-jpa-3/README.md
@@ -11,5 +11,4 @@ This module contains articles about the Java Persistence API (JPA) in Java.
 - [A Guide to MultipleBagFetchException in Hibernate](https://www.baeldung.com/java-hibernate-multiplebagfetchexception)
 - [How to Convert a Hibernate Proxy to a Real Entity Object](https://www.baeldung.com/hibernate-proxy-to-real-entity-object)
 - [How to Return Multiple Entities in JPA Query](https://www.baeldung.com/jpa-return-multiple-entities)
-- [Defining Unique Constraints in JPA](https://www.baeldung.com/jpa-unique-constraints)
 - [Connecting to a Specific Schema in JDBC](https://www.baeldung.com/jdbc-connect-to-schema)
