+++ persistence-modules/java-jpa-3/src/main/resources/META-INF/persistence.xml
@@ -97,23 +97,7 @@
             <property name="hibernate.temp.use_jdbc_metadata_defaults" value="false"/>
         </properties>
     </persistence-unit>
-
-    <persistence-unit name="jpa-unique-constraints">
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <class>com.baeldung.jpa.uniqueconstraints.Person</class>
-        <class>com.baeldung.jpa.uniqueconstraints.Address</class>
-        <exclude-unlisted-classes>true</exclude-unlisted-classes>
-        <properties>
-            <property name="jakarta.persistence.jdbc.driver" value="org.h2.Driver" />
-            <property name="jakarta.persistence.jdbc.url" value="jdbc:h2:mem:test" />
-            <property name="jakarta.persistence.jdbc.user" value="sa" />
-            <property name="jakarta.persistence.jdbc.password" value="" />
-            <property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect" />
-            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
-            <property name="show_sql" value="true" />
-            <property name="hibernate.temp.use_jdbc_metadata_defaults" value="false" />
-        </properties>
-	</persistence-unit>
+    
     <persistence-unit name="jpa-h2-return-multiple-entities">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
         <class>com.baeldung.jpa.returnmultipleentities.Channel</class>
