+++ persistence-modules/java-jpa/src/main/java/com/baeldung/jpa/uniqueconstraints/Address.java
