+++ persistence-modules/java-jpa/src/test/java/com/baeldung/jpa/uniqueconstraints/UniqueColumnIntegrationTest.java
@@ -2,16 +2,17 @@ package com.baeldung.jpa.uniqueconstraints;
 
 import java.util.Optional;
 
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.Persistence;
-
-import org.hibernate.exception.ConstraintViolationException;
+import org.h2.jdbc.JdbcSQLIntegrityConstraintViolationException;
 import org.junit.Assert;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-public class UniqueConstraintIntegrationTest {
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.Persistence;
+
+public class UniqueColumnIntegrationTest {
+    
     private static EntityManagerFactory factory;
     private static EntityManager entityManager;
 
@@ -22,14 +23,43 @@ public class UniqueConstraintIntegrationTest {
     }
 
     @Test
-    public void whenPersistPersonWithSameNumberAndStatus_thenConstraintViolationException() {
+    public void whenPersistPersonWithSameNumber_thenConstraintViolationException() {
         Person person1 = new Person();
-        person1.setPersonNumber(12345L);
-        person1.setIsActive(Boolean.TRUE);
+        person1.setPersonNumber(2000L);
+        person1.setEmail("john.beth@gmail.com");
 
         Person person2 = new Person();
-        person2.setPersonNumber(12345L);
-        person2.setIsActive(Boolean.TRUE);
+        person2.setPersonNumber(2000L);
+        person2.setEmail("anthony.green@gmail.com");
+
+        entityManager.getTransaction().begin();
+        entityManager.persist(person1);
+        entityManager.getTransaction().commit();
+        
+        entityManager.getTransaction().begin();
+        try {
+            entityManager.persist(person2);
+            entityManager.getTransaction().commit();
+            Assert.fail("Should raise an exception - unique key violation");
+        } catch (Exception ex) {
+            Assert.assertTrue(Optional.of(ex)
+                .map(Throwable::getCause)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
+                .isPresent());
+        } finally {
+            entityManager.getTransaction().rollback();
+        }
+    }
+
+    @Test
+    public void whenPersistPersonWithSameEmail_thenConstraintViolationException() {
+        Person person1 = new Person();
+        person1.setPersonNumber(4000L);
+        person1.setEmail("timm.beth@gmail.com");
+
+        Person person2 = new Person();
+        person2.setPersonNumber(3000L);
+        person2.setEmail("timm.beth@gmail.com");
 
         entityManager.getTransaction().begin();
         entityManager.persist(person1);
@@ -43,7 +73,7 @@ public class UniqueConstraintIntegrationTest {
         } catch (Exception ex) {
             Assert.assertTrue(Optional.of(ex)
                 .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
                 .isPresent());
         } finally {
             entityManager.getTransaction().rollback();
@@ -51,45 +81,18 @@ public class UniqueConstraintIntegrationTest {
     }
 
     @Test
-    public void whenPersistPersonWithSameSCodeAndDecode_thenConstraintViolationException() {
+    public void whenPersistPersonWithSameAddress_thenConstraintViolationException() {
         Person person1 = new Person();
-        person1.setDcode("Sec1");
-        person1.setScode("Axybg356");
+        person1.setPersonNumber(5000L);
+        person1.setEmail("chris.beck@gmail.com");
 
-        Person person2 = new Person();
-        person2.setDcode("Sec1");
-        person2.setScode("Axybg356");
-
-        entityManager.getTransaction().begin();
-        entityManager.persist(person1);
-        entityManager.getTransaction().commit();
-
-        entityManager.getTransaction().begin();
-        try {
-            entityManager.persist(person2);
-            entityManager.getTransaction().commit();
-            Assert.fail("Should raise an exception - unique key violation");
-        } catch (Exception ex) {
-            Assert.assertTrue(Optional.of(ex)
-                .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
-                .isPresent());
-        } finally {
-            entityManager.getTransaction().rollback();
-        }
-    }
-
-    @Test
-    public void whenPersistPersonWithSameNumberAndAddress_thenConstraintViolationException() {
         Address address1 = new Address();
-        address1.setStreetAddress("40 Street");
-
-        Person person1 = new Person();
-        person1.setPersonNumber(54321L);
+        address1.setStreetAddress("20 Street");
         person1.setAddress(address1);
 
         Person person2 = new Person();
-        person2.setPersonNumber(99999L);
+        person2.setPersonNumber(6000L);
+        person2.setEmail("mark.jonson@gmail.com");
         person2.setAddress(address1);
 
         entityManager.getTransaction().begin();
@@ -104,7 +107,7 @@ public class UniqueConstraintIntegrationTest {
         } catch (Exception ex) {
             Assert.assertTrue(Optional.of(ex)
                 .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
                 .isPresent());
         } finally {
             entityManager.getTransaction().rollback();
