+++ persistence-modules/java-jpa/src/test/java/com/baeldung/jpa/uniqueconstraints/UniqueConstraintIntegrationTest.java
@@ -2,17 +2,16 @@ package com.baeldung.jpa.uniqueconstraints;
 
 import java.util.Optional;
 
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.Persistence;
-
-import org.hibernate.exception.ConstraintViolationException;
+import org.h2.jdbc.JdbcSQLIntegrityConstraintViolationException;
 import org.junit.Assert;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-public class UniqueColumnIntegrationTest {
-    
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.Persistence;
+
+public class UniqueConstraintIntegrationTest {
     private static EntityManagerFactory factory;
     private static EntityManager entityManager;
 
@@ -23,43 +22,14 @@ public class UniqueColumnIntegrationTest {
     }
 
     @Test
-    public void whenPersistPersonWithSameNumber_thenConstraintViolationException() {
+    public void whenPersistPersonWithSameNumberAndStatus_thenConstraintViolationException() {
         Person person1 = new Person();
-        person1.setPersonNumber(2000L);
-        person1.setEmail("john.beth@gmail.com");
+        person1.setPersonNumber(12345L);
+        person1.setIsActive(Boolean.TRUE);
 
         Person person2 = new Person();
-        person2.setPersonNumber(2000L);
-        person2.setEmail("anthony.green@gmail.com");
-
-        entityManager.getTransaction().begin();
-        entityManager.persist(person1);
-        entityManager.getTransaction().commit();
-        
-        entityManager.getTransaction().begin();
-        try {
-            entityManager.persist(person2);
-            entityManager.getTransaction().commit();
-            Assert.fail("Should raise an exception - unique key violation");
-        } catch (Exception ex) {
-            Assert.assertTrue(Optional.of(ex)
-                .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
-                .isPresent());
-        } finally {
-            entityManager.getTransaction().rollback();
-        }
-    }
-
-    @Test
-    public void whenPersistPersonWithSameEmail_thenConstraintViolationException() {
-        Person person1 = new Person();
-        person1.setPersonNumber(4000L);
-        person1.setEmail("timm.beth@gmail.com");
-
-        Person person2 = new Person();
-        person2.setPersonNumber(3000L);
-        person2.setEmail("timm.beth@gmail.com");
+        person2.setPersonNumber(12345L);
+        person2.setIsActive(Boolean.TRUE);
 
         entityManager.getTransaction().begin();
         entityManager.persist(person1);
@@ -73,7 +43,7 @@ public class UniqueColumnIntegrationTest {
         } catch (Exception ex) {
             Assert.assertTrue(Optional.of(ex)
                 .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
                 .isPresent());
         } finally {
             entityManager.getTransaction().rollback();
@@ -81,18 +51,45 @@ public class UniqueColumnIntegrationTest {
     }
 
     @Test
-    public void whenPersistPersonWithSameAddress_thenConstraintViolationException() {
+    public void whenPersistPersonWithSameSCodeAndDecode_thenConstraintViolationException() {
         Person person1 = new Person();
-        person1.setPersonNumber(5000L);
-        person1.setEmail("chris.beck@gmail.com");
+        person1.setDcode("Sec1");
+        person1.setScode("Axybg356");
 
+        Person person2 = new Person();
+        person2.setDcode("Sec1");
+        person2.setScode("Axybg356");
+
+        entityManager.getTransaction().begin();
+        entityManager.persist(person1);
+        entityManager.getTransaction().commit();
+
+        entityManager.getTransaction().begin();
+        try {
+            entityManager.persist(person2);
+            entityManager.getTransaction().commit();
+            Assert.fail("Should raise an exception - unique key violation");
+        } catch (Exception ex) {
+            Assert.assertTrue(Optional.of(ex)
+                .map(Throwable::getCause)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
+                .isPresent());
+        } finally {
+            entityManager.getTransaction().rollback();
+        }
+    }
+
+    @Test
+    public void whenPersistPersonWithSameNumberAndAddress_thenConstraintViolationException() {
         Address address1 = new Address();
-        address1.setStreetAddress("20 Street");
+        address1.setStreetAddress("40 Street");
+
+        Person person1 = new Person();
+        person1.setPersonNumber(54321L);
         person1.setAddress(address1);
 
         Person person2 = new Person();
-        person2.setPersonNumber(6000L);
-        person2.setEmail("mark.jonson@gmail.com");
+        person2.setPersonNumber(99999L);
         person2.setAddress(address1);
 
         entityManager.getTransaction().begin();
@@ -107,7 +104,7 @@ public class UniqueColumnIntegrationTest {
         } catch (Exception ex) {
             Assert.assertTrue(Optional.of(ex)
                 .map(Throwable::getCause)
-                .filter(x -> x instanceof ConstraintViolationException)
+                .filter(x -> x instanceof JdbcSQLIntegrityConstraintViolationException)
                 .isPresent());
         } finally {
             entityManager.getTransaction().rollback();
